add modal.js
Перенес функционал модалки в файл modal.js
